requirements.txt
# Core dependencies
llama-cpp-python>=0.2.0
numpy>=1.24.0

# Optional: For development
# black>=23.0.0
# pytest>=7.4.0
# mypy>=1.5.0
